Move duplicated logic to trait
